Make `_isShutdown ` volatile
It's read/written by different threads (particularly, it's read by the daemon thread defined in the Pool constructor).

Because access to the variable isn't guarded by a synchronized block or the `volatile` modifier, there isn't a visibility guarantee across threads.

This would be a reasonable explanation for a thread leak I'm experiencing.